added updateContact
