+++ packages/browser-destinations/src/destinations/intercom/utils.ts
@@ -1,0 +1,6 @@
+import dayjs from "dayjs";
+
+//convert from ISO to Unix
+export function convertISOtoUnix(date: string){
+    return dayjs(date).unix()
+}
